requirements.txt
azure-storage-blob>=12.19.0
pyodbc>=5.0.1
Flask>=3.0.0
Flask-CORS>=4.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
